Create TileController  inherited by NormalTileController
